type: use 'unsigned short int' instead of 'int'

+++ vigenere_crypt.hpp
@@ -10,6 +10,6 @@ class VigenereCrypt
     private:
         unsigned int m_count;
         std::string m_key;
-        std::map<char,int> vigenere_table;
+        std::map<char,unsigned short int> vigenere_table;
 };
